Camera movement, shading

+++ src/ray.h
@@ -1,7 +1,7 @@
 #ifndef RAY_H
 #define RAY_H
 
-double ray(int degAngle, double posx, double posy);
+double ray(double degAngle, double posx, double posy);
 void drawRect(double width, double height, double x);
 
 #endif
